requirements.txt
openai
python-dotenv
serpapi